+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - App config for SQL
 - App config for Airbrake
 - App config for Jaeger
+- App config for gRPC client
 
 
 ## [0.6.3] - 2017-12-05